Update to ubuntu.md
In Debian 8.x (32 Bit) /etc/apt/sources.list.d exists, and the sudo add-apt-repository ppa:bqlabs/horus command will execute,
but doesn't create the file, and gives no error.  The only way to get the file built is to execute the command with the -y switch
to assume "yes".  Then, the newly created file "bqlabs-horus-jessie.list" needs to be edited to point to the "trusty" repo, so the
PPA GPG key is automatically imported.

REF:  http://www.webupd8.org/2014/10/how-to-add-launchpad-ppas-in-debian-via.html

THANKS.

Larry Kraemer